requirements/dev.txt
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.28.0
billiard==4.1.0
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect==1.2.0
cssutils==2.7.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
emails==0.6
exceptiongroup==1.1.3
fastapi==0.103.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
mutagen==1.47.0
Naked==0.1.32
orjson==3.9.7
packaging==23.1
passlib==1.7.4
pluggy==1.3.0
premailer==3.10.0
prompt-toolkit==3.0.39
pyasn1==0.5.0
pycryptodome==3.18.0
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
PyJWT==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
rsa==4.9
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
watchfiles==0.20.0
wcwidth==0.2.6
websockets==11.0.3
